Every week is quality week!
New Engineering Principle: Every week is quality week

What do I mean by that statement?

During quality week, we showcased that we are able to make major improvements not just to code, but generally into our product by being radically focused on quality.

All the fixes and improvements from last week made our product more polished, faster and more reliable; Exactly what our customers are asking for.
We also solved a big number of real issues directly for our customers by prioritising bugs reported to us. This work is 100% aligned with our company and engineering values.

So why stop doing this now?

What about our roadmap / features?

Our roadmap is totally dependent on our ability to deliver features on a stable foundation.

It's our job as engineers to prevent the accumulation of bugs and technical debt in order to maintain the trust of our customers and our continued ability to deliver features. We refuse any request to lower our quality standards for the sake of cutting corners.

In our current business context, it's unfortunately clear that we not yet at a good enough level of quality.

This means we need to carry on the critical work we started during quality week. Especially all bugs, stability and security improvements. Our target should be to hit (and surpass) our target quality baseline as soon as possible.

If doing that means postponing work on a new feature – we trust you to make that decision.

Let's make every week quality week from now on!